Update CreateTask button for new style.
